refactor: Clean the code, remove VideoBlockStaticContent class declaration
